==> 02_protocol/10_metadata.Rmd <==
---
title: "Metadata"
---

```{r setup, include=FALSE}

# Do not echo code chunks in rendered document

knitr::opts_chunk$set(echo = FALSE);

# Do not render values returned by code in code chunks

knitr::opts_chunk$set(results = "hide")

# Clear the global environment

rm(list = ls());

# Load the tidyr package for use of the pipe operator

library(tidyr)
library(EMLassemblyline)

```

<!-- Automatically populate a description of the compilation of this file. -->

Generated from R markdown file **./02_protocol/`r knitr::current_input()`**, last compiled on `r Sys.time()`

```{r}

emlalPath <- "./03_incremental/.temp/EML";

# Delete the EML directory to make sure residual files do not hide bugs

unlink(
  x = emlalPath,
  recursive = TRUE
)

# Make sure that target directory for EML exists

dir.create(
  path = emlalPath,
  recursive = TRUE,
  showWarnings = FALSE
);

# Load the incremental metadata from data processing

metaEnv <- readRDS(file = "./03_incremental/.temp/metadata.RData");
metadataNarrative <- readRDS(file = "./03_incremental/meta/metadata_narrative.RData");
metadataPeople <- readRDS(file = "./03_incremental/meta/metadata_people.RData");

```


# Dataset Title

**`r metadataNarrative$title`**


# Abstract

```{r}

# Write the abstract file to the EML directory

write(
  x = metadataNarrative$abstract,
  file = sprintf(
    "%s/abstract.md",
    emlalPath
  )
);

```

`r metadataNarrative$abstract`


# Purpose

`r metadataNarrative$purpose `


# Creators

```{r}

# Define rows of the personnel table for creators

creators <- do.call(
  what = rbind,
  args = lapply(
    X = metadataPeople$creators$elements,
    FUN = function(creator) {
      return(
        data.frame(
          creator$content$givenName,
          creator$content$middleName,
          creator$content$surname,
          creator$content$org,
          creator$content$email,
          creator$content$orcid,
          creator$content$role
        )
      )
    }
  )
);
row.names(creators) <- NULL;
names(creators) <- c(
  "givenName",
  "middleInitial",
  "surName",
  "organizationName",
  "electronicMailAddress",
  "userId",
  "role"
);

personnel <- creators;

creators <- creators[names(creators) != "role"];
creators[is.na(creators)] <- "";
names(creators) <- c(
  "First Name",
  "Middle Initial",
  "Last Name",
  "Organization",
  "e-mail address",
  "ORCID"
);

```

`r knitr::kable(creators) `


# Primary contact 

```{r}

# Define row of the personnel table for the primary contact

contact <- data.frame(
  metadataPeople$contact$givenName,
  metadataPeople$contact$middleName,
  metadataPeople$contact$surname,
  metadataPeople$contact$org,
  metadataPeople$contact$email,
  metadataPeople$contact$orcid,
  metadataPeople$contact$role
);
row.names(contact) <- NULL;
names(contact) <- c(
  "givenName",
  "middleInitial",
  "surName",
  "organizationName",
  "electronicMailAddress",
  "userId",
  "role"
);

personnel <- rbind(personnel, contact);

contact <- contact[names(contact) != "role"];
contact[is.na(contact)] <- ""
names(contact) <- c(
  "First Name",
  "Middle Initial",
  "Last Name",
  "Organization",
  "e-mail address",
  "ORCID"
);

```

`r knitr::kable(contact) `


# Other personnel roles

```{r}

# Define rows of the personnel table for other roles

others <- do.call(
  what = rbind,
  args = lapply(
    X = metadataPeople$others$elements,
    FUN = function(other) {
      return(
        data.frame(
          other$content$givenName,
          other$content$middleName,
          other$content$surname,
          other$content$org,
          other$content$email,
          other$content$orcid,
          other$content$role
        )
      )
    }
  )
);
row.names(others) <- NULL;
names(others) <- c(
  "givenName",
  "middleInitial",
  "surName",
  "organizationName",
  "electronicMailAddress",
  "userId",
  "role"
);

personnel <- rbind(personnel, others);

others[is.na(others)] <- ""
names(others) <- c(
  "First Name",
  "Middle Initial",
  "Last Name",
  "Organization",
  "e-mail address",
  "ORCID",
  "Role"
);

```

`r knitr::kable(others) `


# License

```{r}

# Write the license file to the EML directory

write(
  x = metadataNarrative$intellectualRights,
  file = sprintf(
    "%s/intellectual_rights.txt",
    emlalPath
  )
);

```

`r metadataNarrative$intellectualRights `


# Keywords

```{r}

# Create the keywords table

keywords <- rbind.data.frame(
  list(keyword = "Upper Clark Fork River restoration", keywordThesaurus = ""),
  list(keyword = "NSF DEB LTREB 1655197", keywordThesaurus = ""),
  list(keyword = "Montana", keywordThesaurus = ""),
  list(keyword = "Northwestern United States", keywordThesaurus = ""),
  list(keyword = "Northwestern Forested Mountains Ecoregion", keywordThesaurus = ""),
  list(keyword = "excitation-emission matrix", keywordThesaurus = ""),
  list(keyword = "river water", keywordThesaurus = "NALT"),
  list(keyword = "dissolved organic carbon", keywordThesaurus = "LTER Controlled Vocabulary"),
  list(keyword = "fluorescence", keywordThesaurus = "NALT"),
  list(keyword = "absorbance", keywordThesaurus = "NALT"),
  list(keyword = "spectroscopy", keywordThesaurus = "NALT")
);

# Function to generate a markdown character string with
# the keywords grouped by thesaurus

printKeywords <- function(keywords)
{
  kw <- split(x = keywords, f = keywords$keywordThesaurus);
  names(kw)[names(kw) == ""] <- "No thesaurus";
  kwString <- "";
  for(name in names(kw)) {
    kwString <- paste0(
      kwString,
      "**", name, ":** ",
      paste(kw[[name]]$keyword, collapse = ", "),
      " \n\n"
    );
  }
  return(kwString);
}

```

`r printKeywords(keywords) `

```{r}

# Write the keywords table to the EML directory

write.table(
  x = keywords,
  file = sprintf(
    "%s/keywords.txt",
    emlalPath
  ),
  sep = "\t",
  row.names = FALSE,
  quote = FALSE,
  na = ""
)

```


# Funding of this work

```{r}

# Define rows of the personnel table for funding sources

funding <- do.call(
  what = rbind,
  args = lapply(
    X = metadataPeople$funding$elements,
    FUN = function(fund) {
      return(
        data.frame(
          fund$content$givenName,
          fund$content$middleName,
          fund$content$surname,
          fund$content$org,
          fund$content$email,
          fund$content$orcid,
          fund$content$role,
          fund$content$project,
          fund$content$agency,
          fund$content$awardNumber
        )
      )
    }
  )
);
row.names(funding) <- NULL;
names(funding) <- c(
  "givenName",
  "middleInitial",
  "surName",
  "organizationName",
  "electronicMailAddress",
  "userId",
  "role",
  "projectTitle",
  "fundingAgency",
  "fundingNumber"
);

personnel$projectTitle <- NA;
personnel$fundingAgency <- NA;
personnel$fundingNumber <- NA;
personnel <- rbind(personnel, funding);

funding <- funding[names(funding) != "role"];
funding[is.na(funding)] <- ""
names(funding) <- c(
  "First Name",
  "Middle Initial",
  "Last Name",
  "Organization",
  "e-mail address",
  "ORCID",
  "Project",
  "Funding source",
  "Award number"
);

```

`r knitr::kable(funding) `

```{r}

# Write the personnel table to the EML directory

write.table(
  x = personnel,
  file = sprintf(
    "%s/personnel.txt",
    emlalPath
  ),
  sep = "\t",
  row.names = FALSE,
  quote = FALSE,
  na = ""
)

```


# Timeframe

```{r}

maintenance.description <- "Completed: Updates to these data are not expected"

```

* Begin date: `r format(metaEnv$startTime, format = "%e %B %Y")`, Date code: `r format(metaEnv$startTime, format = "%Y%m%d")`
* End date: `r format(metaEnv$endTime, format = "%e %B %Y")`, Date code: `r format(metaEnv$endTime, format = "%Y%m%d")`
* `r maintenance.description `


# Geographic location

```{r}

# Read the projects sites table

sitesIn <- readxl::read_excel(
  path = "./01_input/meta/2021 Site Table.xlsx",
  sheet = "Table 1",
  range = readxl::cell_limits(
    ul = c(4, 2),
    lr = c(NA, 11)
  ),
  col_names = TRUE,
  col_types = c(
    rep("text", times = 4),
    rep("numeric", times = 6)
  )
);
sites <- as.data.frame(sitesIn[,1:4]);

# Create decimal degress latitude and longitude columns

sites$lat <- 
  sitesIn$LatDegreeNorth + 
  sitesIn$LatMinuteNorth / 60 + 
  sitesIn$LatSecondNorth / 3600;

sites$long <-
  - sitesIn$LongDegreeWest -
  sitesIn$LongMinuteWest / 60 -
  sitesIn$LongSecondWest / 3600;

# Assign row names to site names for indexing

rownames(sites) <- sites$SiteID;

# Filter out the sites covered by this data product

sites <- sites[metaEnv$sites, ];

# Define the table for geographical coverage

geoCoverage <- data.frame(
  geographicDescription = sprintf(
    "Site ID %s, restoration project reach %s: %s (site name: %s)",
    sites$SiteID,
    sites$Reach,
    sites$LocationDesc,
    sites$SiteName
  ),
  northBoundingCoordinate = sites$lat,
  southBoundingCoordinate = sites$lat,
  eastBoundingCoordinate = sites$long,
  westBoundingCoordinate = sites$long
)
colNames <- c(
  "Project Site ID",
  "Restoration reach",
  "Common site name",
  "Location description",
  "Latitude",
  "Longitude"
)

# Define the table of categories associated with sites

sitesCategories <- data.frame(
  attributeName = "siteID",
  code = metaEnv$sites,
  definition = sprintf(
    "Project site number %s",
    metaEnv$sites
  )
)

```

Data are included for the following LTREB project sites. Positive latitudes indicate degrees north and negative longitudes indicate degrees west.

`r knitr::kable(x = sites, row.names = FALSE, col.names = colNames, align = "l")`

```{r}

# Write the geographical coverage table to the EML directory

write.table(
  x = geoCoverage,
  file = sprintf(
    "%s/geographic_coverage.txt",
    emlalPath
  ),
  sep = "\t",
  row.names = FALSE,
  quote = FALSE,
  na = ""
);

```


# Methods

```{r}

write(
  x = metadataNarrative$methods,
  file = sprintf(
    "%s/methods.md",
    emlalPath
  )
);

```

`r metadataNarrative$methods `


# Data tables and other entities

```{r}

tableDoc <- function(longmeta) {
  
  df <- data.frame(
    sprintf(
      fmt = "%s. Property: %s. Entity: %s. Method: %s.",
      longmeta$Desc,
      longmeta$Property,
      longmeta$Entity,
      longmeta$Method
    ),
    longmeta$Units,
    paste(longmeta$EMLUnits, longmeta$EMLTimeFormat),
    longmeta$EMLClass,
    "NA",
    row.names = rownames(longmeta)
  );
  names(df) <- c(
    "Description",
    "Units",
    "EML Units or Time Format",
    "EML Class",
    "Missing value code"
  );
  return(df)
  
}

tableAtt <- function(longmeta, tableName, emlalPath) {

  attributes <- data.frame(
    attributeName = rownames(longmeta),
    attributeDefinition = sprintf(
      fmt = "%s. Property: %s. Entity: %s. Method: %s.",
      longmeta$Desc,
      longmeta$Property,
      longmeta$Entity,
      longmeta$Method
    ),
    class = longmeta$EMLClass,
    unit = longmeta$EMLUnits,
    dateTimeFormatString = longmeta$EMLTimeFormat,
    missingValueCode = "NA",
    missingValueCodeExplanation = "Missing value"
  );

  write.table(
    x = attributes,
    file = sprintf(
      "%s/attributes_%s.txt",
      emlalPath,
      tableName
    ),
    sep = "\t",
    row.names = FALSE,
    quote = FALSE,
    na = ""
  )
  
}

tableNames <- character(length = 0);
tableDescs <- character(length = 0);

```


## Data tables

```{r TableEEMSFiles}

tableName <- metaEnv$tableEEMSFiles.name;
tableNames[length(tableNames) + 1] <- sprintf("%s.csv", tableName);
tableDescs[length(tableDescs) + 1] <- metaEnv$tableEEMSFiles.desc;

```

### *Table name: `r tableNames[length(tableNames)]`*

* **Description:** `r tableDescs[length(tableDescs)]`
* **Table of attribute metadata:**

```{r}

# Convert the column metadata to long form

longmeta <- as.data.frame(t(as.matrix(metaEnv$tableEEMSFiles.meta)));

```

`r knitr::kable(tableDoc(longmeta)) `

```{r}

tableAtt(
  longmeta = longmeta, 
  tableName = tableName, 
  emlalPath = emlalPath
);

# Write the categorical variable tables to the EML directory

write.table(
  x = rbind(
    sitesCategories,
    data.frame(
      attributeName = rep("repID", times = 3),
      code = c("A", "B", "C"),
      definition = c(
        "Replicate sample A",
        "Replicate sample B",
        "Replicate sample C"
      )
    )
  ),
  file = sprintf(
    "%s/catvars_%s.txt",
    emlalPath,
    tableName
  ),
  sep = "\t",
  row.names = FALSE,
  quote = FALSE,
  na = ""
);

```


```{r TableEEMSColumns}

tableName <- metaEnv$tableEEMSCols.name;
tableNames[length(tableNames) + 1] <- sprintf("%s.csv", tableName);
tableDescs[length(tableDescs) + 1] <- metaEnv$tableEEMSCols.desc;

```

### *Table name: `r tableNames[length(tableNames)]`*

* **Description:** `r tableDescs[length(tableDescs)]`
* **Table of attribute metadata:**

```{r}

# Convert the column metadata to long form

longmeta <- as.data.frame(t(as.matrix(metaEnv$tableEEMSCols.meta)));

```

`r knitr::kable(tableDoc(longmeta)) `

```{r}

tableAtt(
  longmeta = longmeta, 
  tableName = tableName, 
  emlalPath = emlalPath
);

```


```{r TableEEMSRows}

tableName <- metaEnv$tableEEMSRows.name;
tableNames[length(tableNames) + 1] <- sprintf("%s.csv", tableName);
tableDescs[length(tableDescs) + 1] <- metaEnv$tableEEMSRows.desc;

```

### *Table name: `r tableNames[length(tableNames)]`*

* **Description:** `r tableDescs[length(tableDescs)]`
* **Table of attribute metadata:**

```{r}

# Convert the column metadata to long form

longmeta <- as.data.frame(t(as.matrix(metaEnv$tableEEMSRows.meta)));

```

`r knitr::kable(tableDoc(longmeta)) `

```{r}

tableAtt(
  longmeta = longmeta, 
  tableName = tableName, 
  emlalPath = emlalPath
);

```


```{r TableAbsorbance}

tableName <- metaEnv$tableAbsorbance.name;
tableNames[length(tableNames) + 1] <- sprintf("%s.csv", tableName);
tableDescs[length(tableDescs) + 1] <- metaEnv$tableAbsorbance.desc;

```

### *Table name: `r tableNames[length(tableNames)]`*

* **Description:** `r tableDescs[length(tableDescs)]`
* **Table of attribute metadata:**

```{r}

# Convert the column metadata to long form

longmeta <- as.data.frame(t(as.matrix(metaEnv$tableAbsorbance.meta)));

```

`r knitr::kable(tableDoc(longmeta)) `

```{r}

tableAtt(
  longmeta = longmeta, 
  tableName = tableName, 
  emlalPath = emlalPath
);

# Write the categorical variable tables to the EML directory

write.table(
  x = rbind(
    sitesCategories,
    data.frame(
      attributeName = rep("repID", times = 3),
      code = c("A", "B", "C"),
      definition = c(
        "Replicate sample A",
        "Replicate sample B",
        "Replicate sample C"
      )
    )
  ),
  file = sprintf(
    "%s/catvars_%s.txt",
    emlalPath,
    tableName
  ),
  sep = "\t",
  row.names = FALSE,
  quote = FALSE,
  na = ""
);

```


```{r TableSummaryMetrics}

tableName <- metaEnv$tableSummaryMetrics.name;
tableNames[length(tableNames) + 1] <- sprintf("%s.csv", tableName);
tableDescs[length(tableDescs) + 1] <- metaEnv$tableSummaryMetrics.desc;

```

### *Table name: `r tableNames[length(tableNames)]`*

* **Description:** `r tableDescs[length(tableDescs)]`
* **Table of attribute metadata:**

```{r}

# Convert the column metadata to long form

longmeta <- as.data.frame(t(as.matrix(metaEnv$tableSummaryMetrics.meta)));

```

`r knitr::kable(tableDoc(longmeta)) `

```{r}

tableAtt(
  longmeta = longmeta, 
  tableName = tableName, 
  emlalPath = emlalPath
);

# Write the categorical variable tables to the EML directory

write.table(
  x = rbind(
    sitesCategories,
    data.frame(
      attributeName = rep("repID", times = 3),
      code = c("A", "B", "C"),
      definition = c(
        "Replicate sample A",
        "Replicate sample B",
        "Replicate sample C"
      )
    )
  ),
  file = sprintf(
    "%s/catvars_%s.txt",
    emlalPath,
    tableName
  ),
  sep = "\t",
  row.names = FALSE,
  quote = FALSE,
  na = ""
);

```


## Other data entities

```{r results = 'asis'}

otherNames <- c(
  sprintf("%s_matrices.zip", metaEnv$tableEEMSFiles.name),
  sprintf("%s_vis.pdf", metaEnv$tableSummaryMetrics.name),
  "01_input.zip",
  "02_protocol.zip",
  "03_incremental.zip"
);
otherDescs <- c(
  sprintf(
    fmt = paste(
      "Compressed zip file with fluorometric excitation-emission matrices.",
      "Matrix columns represent excitation wavelengths and",
      "matrix rows represent emission wavelengths",
      "(see Methods for details).",
      "Files are indexed by sampling date, site ID, and replicate ID",
      "in the table %s.csv."
    ),
    metaEnv$tableEEMSFiles.name
  ),
  "Visualization of the absorbance and fluorescence summary metrics over time.",
  paste(
    "Compressed data pipeline folder containing project default metadata.",
    "See methods for details."
  ),
  paste(
    "Compressed data pipeline folder containing R markdown and R scripts",
    "for processing data and metadata.",
    "See methods for details."
  ),
  paste(
    "Compressed data pipeline folder containing detailed notes",
    "and incremental data associated with generation of the data product.",
    "See methods for details."
  )
);

cat(
  paste(
    "\n\n### *File name:", 
    otherNames,
    "*",
    "\n\n* **Description:**", 
    otherDescs
  )
);

```


# Zip the pipeline

```{r}
# Customized zip function for capturing output (doctored version of utils::zip)

customZip <- function (
  zipfile, files, flags = "-r9X", 
  extras = "", zip = Sys.getenv("R_ZIPCMD", "zip"), 
  stdout = TRUE, stderr = TRUE
) 
{
  if (missing(flags) && (!is.character(files) || !length(files))) 
    stop("'files' must be a character vector specifying one or more filepaths")
  if (!is.character(zip) || length(zip) != 1L || !nzchar(zip)) 
    stop("argument 'zip' must be a non-empty character string")
  args <- c(flags, shQuote(path.expand(zipfile)), shQuote(files), 
            extras)
  if (sum(nchar(c(args, Sys.getenv()))) + length(args) > 8000) {
    args <- c(flags, "-@", shQuote(path.expand(zipfile)), 
              extras)
    input <- files
  }
  else input <- NULL
  if (.Platform$OS.type == "windows") 
    invisible(system2(zip, args, input = input, invisible = TRUE, stdout = stdout, stderr = stderr))
  else invisible(system2(zip, args, input = input, stdout = stdout, stderr = stderr))
}

```

Create the pipeline zip files. Warnings with status 12 are normal if the zip file already exists and no files need to be freshened.

## 01_input.zip

```{r results = "asis"}

exclusions <- c(
  "./01_input/temp/*", 
  "./01_input/.temp/*"
)

if (length(exclusions) > 0) {
  cat("*This pipeline directory is zipped with the following paths excluded:*\n\n")
  cat(
    paste(
      gsub("[*]", "\\\\*", exclusions),
      collapse = "  \n"
    )
  )
} 

```

```{r results = "asis"}

output <- customZip(
  zipfile = "./04_product/01_input.zip",
  files = c(
    "./01_input"
  ),
  flags = ifelse(
    file.exists("./04_product/01_input.zip"), 
    "-r9u", 
    "-r9"
  ),
  extras = paste(
    "-x",
    paste(shQuote(exclusions), collapse = " ")
  )
)

if (length(output) > 0) {
  cat("*Output from call to zip:* \n\n")
  cat(
    paste(output, collapse = "  \n")
  )
} 

```

## 02_protocol.zip:

```{r results = "asis"}

exclusions <- c(
  "./02_protocol/temp/*", 
  "./02_protocol/.temp/*",
  "./02_protocol/.Rproj.user/*",
  "./02_protocol/.Rhistory"
)

if (length(exclusions) > 0) {
  cat("*This pipeline directory is zipped with the following paths excluded:*\n\n")
  cat(
    paste(
      gsub("[*]", "\\\\*", exclusions),
      collapse = "  \n"
    )
  )
} 

```

```{r results = "asis"}

output <- customZip(
  zipfile = "./04_product/02_protocol.zip",
  files = c(
    "./02_protocol"
  ),
  flags = ifelse(
    file.exists("./04_product/02_protocol.zip"), 
    "-r9u", 
    "-r9"
  ),
  extras = paste(
    "-x",
    paste(shQuote(exclusions), collapse = " ")
  )
)

if (length(output) > 0) {
  cat("*Output from call to zip:* \n\n")
  cat(
    paste(output, collapse = "  \n")
  )
} 

```

## 03_incremental.zip:

```{r results = "asis"}

exclusions <- c(
  "./03_incremental/temp/*", 
  "./03_incremental/.temp/*",
  "./03_incremental/metadata_summary.html"
)

if (length(exclusions) > 0) {
  cat("*This pipeline directory is zipped with the following paths excluded:*\n\n")
  cat(
    paste(
      gsub("[*]", "\\\\*", exclusions),
      collapse = "  \n"
    )
  )
} 

```

```{r results = "asis"}

output <- customZip(
  zipfile = "./04_product/03_incremental.zip",
  files = c(
    "./03_incremental"
  ),
  flags = ifelse(
    file.exists("./04_product/03_incremental.zip"), 
    "-r9u", 
    "-r9"
  ),
  extras = paste(
    "-x",
    paste(shQuote(exclusions), collapse = " ")
  )
)

if (length(output) > 0) {
  cat("*Output from call to zip:* \n\n")
  cat(
    paste(output, collapse = "  \n")
  )
} 

```


# Test build of EML file

```{r}

if (exists("customUnits")) {
  
  write.table(
    x = customUnits,
    file = sprintf(
      "%s/custom_units.txt",
      emlalPath
    ),
    sep = "\t",
    row.names = FALSE,
    quote = FALSE,
    na = ""
  )
  
}

customUnitsOut <- function() {
  if(exists("customUnits")) {
    return(
      c(
        "Custom units table: <BR><BR>",
        knitr::kable(customUnits, row.names = FALSE, align = "l")
      )
    );
  }
}

```

`r if(exists("customUnits")) { "Custom units table: <BR><BR>"} `
`r if(exists("customUnits")) { knitr::kable(customUnits, row.names = FALSE, align = "l") } `

Output from the call to EMLassemblyline::make_eml:

```{r}

package.id <- "edi.1182.1"

emlFilePath <- sprintf(
  "./04_product/%s.xml",
  package.id
);

makeEML <- function() {
  make_eml(
    path = emlalPath,
    data.path = "./04_product",
    eml.path = "./04_product", 
    dataset.title = metadataNarrative$title, 
    temporal.coverage = c(
      format(metaEnv$startTime, "%Y-%m-%d"), 
      format(metaEnv$endTime, "%Y-%m-%d")
    ), 
    maintenance.description = maintenance.description, 
    data.table = tableNames, 
    data.table.description = tableDescs,
    data.table.quote.character = rep("\"", times = length(tableNames)),
    other.entity = otherNames,
    other.entity.description = otherDescs,
    user.id = "UCFRResearch",
    user.domain = "EDI", 
    package.id = package.id
  );
  
  # Read the eml file
  
  emlxml <- xml2::read_xml(x = emlFilePath);
  
  # Replace intellectual rights with markdown element
  
  intRights <- xml2::xml_find_first(emlxml, ".//dataset/intellectualRights");
  
  markdown <- intRights %>% 
    xml2::xml_replace("intellectualRights") %>% 
    xml2::xml_add_child("markdown");
  xml2::xml_text(markdown) <- gsub(
    pattern = "\n\n", 
    replacement = "\r\n\r\n", 
    metadataNarrative$intellectualRights
  );
  
  # Add the purpose as a markdown element
  
  coverage <- xml2::xml_find_first(emlxml, ".//dataset/coverage");
  purpose <- coverage %>% xml2::xml_add_sibling("purpose") %>% xml2::xml_add_child("markdown");
  xml2::xml_text(purpose) <- metadataNarrative$purpose;
  
  # Write the xml
  
  xml2::write_xml(emlxml, file = emlFilePath);
  
}

makeEML();

```

Output from the call to EMLassemblyline::issues: 

```{r}

EMLassemblyline::issues();

```


# Summary of R session used to compile this Rmarkdown

```{r results = "markup"}

print(sessionInfo()) 

```
